Refresh grids on equipment code change
Removed 'mounted' state and related lines of code since it was always
set to false after the first re-renders. This would make it impossible
to make new grid requests.
It was not clear why it was there, might have been related to memory
leak warning (which react18 does not output).