[EAM-2446] Add multiple autocomplete support (WIP); Update createOnChangeHandler helper and add one to handle object updates
The multi autocomplete is considered WIP, but it was built as an extension of
the current EAMAutocomplete, so it should not be hard to merge the two.

The createOnChangeHandler now resets 'desc' and 'org' values
when only the value corresponding to 'valueKey' was updated.
Also, it now allows additional arguments to be passed to the
updating function.

Additionally, there is now a 'createOnChangeHandlerObjectUpdate'
that can be used when the updating function expects an object as its
first argument (as opposed to a key and a value arguments).