[COCO-282] Add block and disable scroll up to grid
By adding block to the API we can now explicitly block the grid from
user interaction. A use case is while a cell is saving a new value to
the server according to the user input.

The disable scroll up allows to avoid the scrolling up to the first row
of the grid whenever the rows update. This is specially useful when
the update was not meaningful to an extent that the grid needs to be
scrolled up.

Also, added a check for whether cache is defined before checking the
height. This is helpful mostly for development so that the app doe s
not crash on a live reload.